refactor: 전체 디렉터리 구조 변경
