+++ src/main/java/lotto/model/LottoGenerator.java
@@ -1,8 +1,9 @@
-package lotto;
+package lotto.model;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.ArrayList;
 import java.util.List;
+import lotto.model.Lotto;
 
 public class LottoGenerator {
