+++ src/main/java/lotto/model/LottoResultChecker.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.model;
 
 import java.util.HashMap;
 import java.util.List;
