+++ src/main/java/lotto/model/LottoStatistics.java
@@ -1,6 +1,7 @@
-package lotto;
+package lotto.model;
 
 import java.util.Map;
+import lotto.model.WinningRank;
 
 public class LottoStatistics {
