+++ src/main/java/lotto/validation/InputValidate.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.validation;
 
 import java.util.HashSet;
 import java.util.List;
