+++ src/main/java/lotto/view/InputHandler.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.view;
 
 import camp.nextstep.edu.missionutils.Console;
